project/apps/spa_app/api/views.py
from rest_framework import viewsets, mixins
from rest_framework.permissions import IsAuthenticated
#from project.apps.twyla.api import serializers
#from project.apps.twyla.models import Book, Rate
from django.db.models import Q
from project.pagination import BasePaginator
