requirements.txt
args==0.1.0
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
clint==0.5.1
Django==2.0.3
django-appconf==1.0.2
django-compressor==2.2
django-jinja==2.2.1
django-jinja2==0.1
django-webpack-loader==0.6.0
djangorestframework==3.7.7
factory-boy==2.9.2
fake-factory==0.7.4
Faker==0.8.11
idna==2.6
Jinja2==2.10
linecache2==1.0.0
MarkupSafe==1.0
pkginfo==1.4.1
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pytest==3.4.2
python-dateutil==2.6.1
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
requests-toolbelt==0.8.0
rjsmin==1.0.12
six==1.11.0
text-unidecode==1.2
traceback2==1.4.0
twine==1.7.4
unittest2==1.1.0
urllib3==1.22
django-extensions==2.0.6
python-stdnum==1.8.1
